pyproject.toml
[tool.poetry]
name = "django-third-party-service"
version = "0"
description = ""
authors = ["ali <alysfryly075@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.7"
djangorestframework = "^3.15.2"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
